Implement service worker update detection and streamline imports in layout components
